Add files via upload
So in this project we as complete beginners wished to make a website regarding all the different events/workshops organized by clubs of our iit roorkee. 
The basic premise of this website was for it to have a user login through which students can view the notifications uploaded by the clubs and an Admin Login through which the club admin can upload new notifications for everyone to view .
After all that tedious work and learning we eventually came up with a semi-working model of this website  where the admin login and notification upload works well but the user cannot view the notifications due to an error in the databases.